Refactor: PD-246 나머지 회원가입 필드 분리
